nominatim: move DBConnection class into its own file
